requirements.txt
django>=6.0,<6.1
djangorestframework>=3.16,<3.17
drf-yasg>=1.21,<1.22
